extract-load/requirements.txt
boto3==1.40.72
botocore==1.40.72
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
dlt==1.18.2
dotenv==0.9.9
duckdb==1.4.2
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
giturlparse==0.14.0
humanize==4.14.0
idna==3.11
jmespath==1.0.1
jsonpath-ng==1.7.0
markdown-it-py==4.0.0
mdurl==0.1.2
orjson==3.11.4
packaging==25.0
pathvalidate==3.3.1
pendulum==3.1.0
pluggy==1.6.0
ply==3.11
psycopg2-binary==2.9.11
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requirements-parser==0.13.0
rich==14.2.0
rich-argparse==1.7.2
s3transfer==0.14.0
semver==3.0.4
setuptools==80.9.0
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
sqlglot==27.29.0
tenacity==9.1.2
tomlkit==0.13.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
